Neke Save

+++ ParkingServis/Entiteti/Fizicko_lice.cs
@@ -10,15 +10,15 @@ namespace ParkingServis.Entiteti
     {
         //public virtual int ID { get; protected set; }
 
-        public virtual int JMBG { get; set; }
-        public virtual int LK { get; set; }
+        public virtual long JMBG { get; set; }
+        public virtual long LK { get; set; }
         public virtual String mesto_izdavanja_lk { get; set; }
         public virtual String licno_ime { get; set; }
         public virtual String ime_roditelja { get; set; }
         public virtual String prezime { get; set; }
         public virtual String adresa { get; set; }
         public virtual String prebivaliste { get; set; }
-        public virtual int broj_vozacke { get; set; }
+        public virtual long broj_vozacke { get; set; }
 
     public Fizicko_lice() : base()
         {
